runner/requirements.txt
nats-py==2.6.0

